requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.2.1
antlr4-python3-runtime==4.8
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.1.0
black==21.4b2
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.1.0
cycler==0.11.0
#detectron2 @ git+https://github.com/facebookresearch/detectron2.git@710e7795d0eeadf9def0e7ef957eea13532e34cf
docker-pycreds==0.4.0
fairscale==0.4.8
fire==0.4.0
fonttools==4.36.0
frozenlist==1.3.1
fsspec==2022.7.1
future==0.18.2
fvcore==0.1.5.post20220512
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.11.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
hydra-core==1.0.5
idna==3.3
imageio==2.21.1
importlib-metadata==3.10.1
iopath==0.1.9
joblib==1.1.0
kiwisolver==1.4.4
loguru==0.6.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
#MinkowskiEngine==0.5.4
multidict==6.0.2
mypy-extensions==0.4.3
networkx==2.8.6
ninja==1.10.2.3
numpy==1.23.2
oauthlib==3.2.0
omegaconf==2.0.6
opencv-python-headless==4.6.0.66
packaging==21.3
pathspec==0.9.0
pathtools==0.1.2
Pillow==9.2.0
platformdirs==2.5.2
plyfile==0.7.4
#pointnet2==0.0.0
portalocker==2.5.1
promise==2.3
protobuf==3.19.4
psutil==5.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pyDeprecate==0.3.2
pydot==1.4.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytorch-lightning==1.7.2
pyviz3d==0.2.28
PyWavelets==1.3.0
PyYAML==5.4.1
qudida==0.0.4
regex==2022.8.17
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.0
sentry-sdk==1.9.5
setproctitle==1.3.2
shortuuid==1.0.9
six==1.16.0
sklearn==0.0
smmap==5.0.0
tabulate==0.8.10
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2022.8.12
timm==0.6.7
toml==0.10.2
tomli==2.0.1
#torch==1.12.1
torch-scatter==2.0.9
torchmetrics==0.9.3
#torchvision==0.13.1
tqdm==4.64.0
trimesh==3.14.0
typing_extensions==4.3.0
urllib3==1.26.12
volumentations==0.1.8
wandb==0.13.2
Werkzeug==2.2.2
yacs==0.1.8
yarl==1.8.1
zipp==3.8.1
